+++ src/lib/components/index.ts
@@ -5,4 +5,3 @@ export * from "./main";
 export * from "./minimized-dock";
 export * from "./ui/icons";
 export * from "./updater";
-export * from "./verification-status";